requirements.txt
flagembedding==1.3.5
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_core==0.3.72
langchain_groq==0.3.6
langchain_huggingface==0.3.1
langchain_litellm==0.2.2
langchain_ollama==0.3.6
langchain_text_splitters==0.3.9
litellm==1.74.9
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
streamlit==1.45.1